разбивка на модули
